requirements.txt
Django
djangorestframework
psycopg2-binary
Pillow
python-decouple
django-filter
djangorestframework_simplejwt
python-slugify
drf-yasg
gunicorn
django-cors-headers
whitenoise